Refactor: Replace argument parsing algorithm
Replace the argument parsing algorithm for the `doc` and `help` commands from
a `switch` based approach to an array lookup. At some point this should be
refined so the table lookup is a function in a separate class, but this is
being deferred until Sherpa evovles and command/argument complexity is
better defined.

Resolves: #8
See also: N/a